requirements.txt
# Web Scraping Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
selenium>=4.15.0
selenium-stealth>=1.0.6
openpyxl>=3.1.0
lxml>=4.9.0

# Backend Dependencies
fastapi
uvicorn
python-dotenv
pydantic-settings>=2.1.0

# AI / LLM Dependencies
groq>=0.4.0

# Forecasting Dependencies
prophet>=1.1.0
torch>=2.0.0
chronos-forecasting
numpy>=1.24.0
scikit-learn>=1.3.0

# Testing Dependencies
pytest
pytest-asyncio
httpx
flake8

